fix(e2e): make the nightly e2e suites green
The harness could not reliably kill wrangler: walking the process tree with
pgrep left a window in which wrangler respawned workerd, and the orphan kept
the test's fixed port. Spawn wrangler detached and SIGKILL the process group
instead, sharing one helper across the Think e2e files.

With the harness fixed, three real problems surfaced: Think's in-flight-turn
probes only counted cf_agents_runs although chat turns run on Tasks; the
post-persist failure test asserted onError is never called, which Tasks now
does; and a Think turn interrupted before its first chunk was classified as a
continuation rather than a retry, so agent-tool parents sealed the child as
errored. The ai-chat work-budget agent banked no work under the stream-derived
work marker and now flushes a full segment.

Also fixes the broken shell quoting in the Playwright version step, which made
the cache key empty in every run.
