+++ packages/agents/src/e2e-tests/fiber-eviction.test.ts
@@ -81,28 +81,6 @@ function killProcessOnPort(port: number): void {
   }
 }
 
-function killProcessTree(pid: number): void {
-  let children: number[] = [];
-  try {
-    children = execSync(`pgrep -P ${pid} 2>/dev/null || true`)
-      .toString()
-      .trim()
-      .split("\n")
-      .filter(Boolean)
-      .map(Number);
-  } catch {
-    // pgrep may be unavailable; killing the parent is still useful.
-  }
-  for (const childPid of children) {
-    killProcessTree(childPid);
-  }
-  try {
-    process.kill(pid, "SIGKILL");
-  } catch {
-    // Already dead
-  }
-}
-
 function startWrangler(): ChildProcess {
   const configPath = path.join(__dirname, "wrangler.jsonc");
   const child = spawn(
@@ -122,6 +100,9 @@ function startWrangler(): ChildProcess {
     {
       cwd: __dirname,
       stdio: ["pipe", "pipe", "pipe"],
+      // A process-group leader, so killProcess() can take down wrangler and
+      // every workerd it spawns in one signal.
+      detached: true,
       env: { ...process.env, NODE_ENV: "test" }
     }
   );
@@ -178,7 +159,20 @@ function killProcess(child: ChildProcess): Promise<void> {
       clearTimeout(fallback);
       resolve();
     });
-    killProcessTree(child.pid);
+    // The child is a process-group leader (spawned detached): one signal
+    // takes down npm exec, wrangler, esbuild, and every workerd. Walking the
+    // tree with pgrep and killing children first leaves a window in which
+    // wrangler respawns workerd; that orphan keeps the port and the restart
+    // fails with "Address already in use".
+    try {
+      process.kill(-child.pid, "SIGKILL");
+    } catch {
+      try {
+        process.kill(child.pid, "SIGKILL");
+      } catch {
+        // Already dead.
+      }
+    }
   });
 }
