+++ packages/think/src/e2e-tests/chat-recovery.test.ts
@@ -8,7 +8,8 @@
  * 5. Verify: onChatRecovery fired, partial text persisted, fiber row cleaned up
  */
 import { describe, it, expect, afterEach, beforeEach } from "vitest";
-import { spawn, execSync, type ChildProcess } from "node:child_process";
+import { spawn, type ChildProcess } from "node:child_process";
+import { killProcess, killProcessOnPort } from "./wrangler-process";
 import { setDefaultAutoSelectFamily } from "node:net";
 import "./harden-net";
 import path from "node:path";
@@ -54,49 +55,6 @@ function sleep(ms: number): Promise<void> {
   return new Promise((resolve) => setTimeout(resolve, ms));
 }
 
-function killProcessOnPort(port: number): void {
-  try {
-    const output = execSync(
-      `lsof -tiTCP:${port} -sTCP:LISTEN 2>/dev/null || true`
-    )
-      .toString()
-      .trim();
-    if (output) {
-      for (const pid of output.split("\n").filter(Boolean)) {
-        try {
-          process.kill(Number(pid), "SIGKILL");
-        } catch {
-          // Already dead
-        }
-      }
-    }
-  } catch {
-    // lsof not available
-  }
-}
-
-function killProcessTree(pid: number): void {
-  let children: number[] = [];
-  try {
-    children = execSync(`pgrep -P ${pid} 2>/dev/null || true`)
-      .toString()
-      .trim()
-      .split("\n")
-      .filter(Boolean)
-      .map(Number);
-  } catch {
-    // pgrep may be unavailable; killing the parent is still useful.
-  }
-  for (const childPid of children) {
-    killProcessTree(childPid);
-  }
-  try {
-    process.kill(pid, "SIGKILL");
-  } catch {
-    // Already dead
-  }
-}
-
 function startWrangler(): ChildProcess {
   const configPath = path.join(__dirname, "wrangler.jsonc");
   const child = spawn(
@@ -116,6 +74,9 @@ function startWrangler(): ChildProcess {
     {
       cwd: __dirname,
       stdio: ["pipe", "pipe", "pipe"],
+      // A process-group leader, so killProcess() can take down wrangler and
+      // every workerd it spawns in one signal.
+      detached: true,
       env: { ...process.env, NODE_ENV: "test" }
     }
   );
@@ -157,21 +118,6 @@ async function waitForPortFree(maxAttempts = 30, delayMs = 500): Promise<void> {
     await sleep(delayMs);
   }
   throw new Error(`Port ${PORT} did not free in time`);
-}
-
-function killProcess(child: ChildProcess): Promise<void> {
-  return new Promise((resolve) => {
-    if (!child.pid) {
-      resolve();
-      return;
-    }
-    const fallback = setTimeout(resolve, 3000);
-    child.on("exit", () => {
-      clearTimeout(fallback);
-      resolve();
-    });
-    killProcessTree(child.pid);
-  });
 }
 
 async function restartWrangler(child: ChildProcess): Promise<ChildProcess> {
@@ -557,9 +503,16 @@ describe("Think chat recovery e2e", () => {
     expect(status.messageCount).toBe(1);
     expect(status.assistantMessages).toBe(0);
 
-    // This documents the observability gap from the report: the request catch
-    // broadcasts a chat error frame, but it does not route through onError.
-    expect((await callAgent("getOnErrorLog")) as string[]).toEqual([]);
+    // The turn runs as a Tasks run; its failure is reported through `onError`
+    // (the observability gap from the original report is closed) as well as
+    // through the chat error frame and `onChatError` below.
+    const onErrorLog = await pollUntil(
+      "onError hook",
+      () => callAgent("getOnErrorLog") as Promise<string[]>,
+      (log) => log.some((entry) => entry.includes("forced beforeTurn failure")),
+      { attempts: 10, delayMs: 250 }
+    );
+    expect(onErrorLog).toContain("forced beforeTurn failure");
     const chatErrorLog = await pollUntil(
       "chat error hook",
       () => callAgent("getOnChatErrorLog") as Promise<string[]>,
