+++ packages/think/src/e2e-tests/messenger-recovery.test.ts
@@ -14,7 +14,8 @@
  * adapter that records into agent SQL.
  */
 import { describe, it, expect, afterEach, beforeEach } from "vitest";
-import { spawn, execSync, type ChildProcess } from "node:child_process";
+import { spawn, type ChildProcess } from "node:child_process";
+import { killProcess, killProcessOnPort } from "./wrangler-process";
 import { setDefaultAutoSelectFamily } from "node:net";
 import "./harden-net";
 import path from "node:path";
@@ -32,49 +33,6 @@ const PERSIST_DIR = path.join(__dirname, ".wrangler-think-messenger-e2e-state");
 
 function sleep(ms: number): Promise<void> {
   return new Promise((resolve) => setTimeout(resolve, ms));
-}
-
-function killProcessOnPort(port: number): void {
-  try {
-    const output = execSync(
-      `lsof -tiTCP:${port} -sTCP:LISTEN 2>/dev/null || true`
-    )
-      .toString()
-      .trim();
-    if (output) {
-      for (const pid of output.split("\n").filter(Boolean)) {
-        try {
-          process.kill(Number(pid), "SIGKILL");
-        } catch {
-          // Already dead
-        }
-      }
-    }
-  } catch {
-    // lsof not available
-  }
-}
-
-function killProcessTree(pid: number): void {
-  let children: number[] = [];
-  try {
-    children = execSync(`pgrep -P ${pid} 2>/dev/null || true`)
-      .toString()
-      .trim()
-      .split("\n")
-      .filter(Boolean)
-      .map(Number);
-  } catch {
-    // pgrep may be unavailable; killing the parent is still useful.
-  }
-  for (const childPid of children) {
-    killProcessTree(childPid);
-  }
-  try {
-    process.kill(pid, "SIGKILL");
-  } catch {
-    // Already dead
-  }
 }
 
 function startWrangler(): ChildProcess {
@@ -96,6 +54,9 @@ function startWrangler(): ChildProcess {
     {
       cwd: __dirname,
       stdio: ["pipe", "pipe", "pipe"],
+      // A process-group leader, so killProcess() can take down wrangler and
+      // every workerd it spawns in one signal.
+      detached: true,
       env: { ...process.env, NODE_ENV: "test" }
     }
   );
@@ -137,21 +98,6 @@ async function waitForPortFree(maxAttempts = 30, delayMs = 500): Promise<void> {
     await sleep(delayMs);
   }
   throw new Error(`Port ${PORT} did not free in time`);
-}
-
-function killProcess(child: ChildProcess): Promise<void> {
-  return new Promise((resolve) => {
-    if (!child.pid) {
-      resolve();
-      return;
-    }
-    const fallback = setTimeout(resolve, 3000);
-    child.on("exit", () => {
-      clearTimeout(fallback);
-      resolve();
-    });
-    killProcessTree(child.pid);
-  });
 }
 
 async function restartWrangler(child: ChildProcess): Promise<ChildProcess> {
