+++ packages/think/src/e2e-tests/step-prompt-structured.test.ts
@@ -17,7 +17,8 @@
  *   OPENAI_API_KEY=sk-... ANTHROPIC_API_KEY=sk-ant-... pnpm run test:e2e
  */
 import { describe, it, expect, afterAll, beforeAll } from "vitest";
-import { spawn, execSync, type ChildProcess } from "node:child_process";
+import { spawn, type ChildProcess } from "node:child_process";
+import { killProcess, killProcessOnPort } from "./wrangler-process";
 import { setDefaultAutoSelectFamily } from "node:net";
 import "./harden-net";
 import path from "node:path";
@@ -52,45 +53,6 @@ function sleep(ms: number): Promise<void> {
   return new Promise((resolve) => setTimeout(resolve, ms));
 }
 
-function killProcessOnPort(port: number): void {
-  try {
-    const output = execSync(`lsof -ti tcp:${port} 2>/dev/null || true`)
-      .toString()
-      .trim();
-    if (output) {
-      for (const pid of output.split("\n").filter(Boolean)) {
-        try {
-          process.kill(Number(pid), "SIGKILL");
-        } catch {
-          // already exited
-        }
-      }
-    }
-  } catch {
-    // ignore
-  }
-}
-
-function killProcessTree(pid: number): void {
-  let children: number[] = [];
-  try {
-    children = execSync(`pgrep -P ${pid} 2>/dev/null || true`)
-      .toString()
-      .trim()
-      .split("\n")
-      .filter(Boolean)
-      .map(Number);
-  } catch {
-    // pgrep may be unavailable
-  }
-  for (const childPid of children) killProcessTree(childPid);
-  try {
-    process.kill(pid, "SIGKILL");
-  } catch {
-    // already dead
-  }
-}
-
 function startWrangler(): ChildProcess {
   const configPath = path.join(__dirname, "wrangler.jsonc");
   const args = [
@@ -117,6 +79,9 @@ function startWrangler(): ChildProcess {
   const child = spawn("npx", args, {
     cwd: __dirname,
     stdio: ["pipe", "pipe", "pipe"],
+    // A process-group leader, so killProcess() can take down wrangler and
+    // every workerd it spawns in one signal.
+    detached: true,
     env: { ...process.env, NODE_ENV: "test" }
   });
 
@@ -144,21 +109,6 @@ async function waitForReady(maxAttempts = 60, delayMs = 1000): Promise<void> {
     await sleep(delayMs);
   }
   throw new Error(`Wrangler did not start within ${maxAttempts * delayMs}ms`);
-}
-
-function killProcess(child: ChildProcess): Promise<void> {
-  return new Promise((resolve) => {
-    if (!child.pid) {
-      resolve();
-      return;
-    }
-    const fallback = setTimeout(resolve, 3000);
-    child.on("exit", () => {
-      clearTimeout(fallback);
-      resolve();
-    });
-    killProcessTree(child.pid);
-  });
 }
 
 /** Call a @callable method on the agent via WebSocket RPC. */
